requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas>=2.2.0
numpy>=1.26.0
streamlit==1.28.1
plotly==5.17.0
webdriver-manager==4.0.1
lxml>=5.0.0
feedparser==6.0.10
schedule==1.2.0
python-dateutil==2.8.2
jinja2==3.1.2
markdown==3.5.1
aiohttp==3.9.1 